fix(functional): split provider test helpers
