+++ tests/functional/providers/helpers_test.go
@@ -10,7 +10,6 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"reflect"
 	"strconv"
@@ -20,7 +19,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/portpowered/infinite-you/internal/testutil"
 	platformhttpserver "github.com/portpowered/infinite-you/pkg/platform/httpserver"
 	platformprocess "github.com/portpowered/infinite-you/pkg/platform/process"
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
@@ -992,639 +990,4 @@ func (fixture *ProcessFixture) validateSessionTopology() error {
 
 func (fixture *ProcessFixture) RuntimeLogDir() string {
 	return fixture.runtimeLogs
-}
-
-func FindRuntimeLogRecord(
-	t *testing.T,
-	fixture *ProcessFixture,
-	workDir string,
-	exitCode int,
-) map[string]any {
-	t.Helper()
-	var found map[string]any
-	err := filepath.WalkDir(fixture.runtimeLogs, func(path string, entry fs.DirEntry, walkErr error) error {
-		if walkErr != nil {
-			return walkErr
-		}
-		if entry.IsDir() || filepath.Ext(path) != ".log" {
-			return nil
-		}
-		file, err := os.Open(path)
-		if err != nil {
-			return err
-		}
-		defer file.Close()
-		decoder := json.NewDecoder(file)
-		for {
-			var record map[string]any
-			if err := decoder.Decode(&record); err != nil {
-				if err == io.EOF {
-					break
-				}
-				return err
-			}
-			if record["event_name"] != commandRunnerCompletedLogEvent {
-				continue
-			}
-			if got, ok := record["working_dir"].(string); ok && workDir != "" && got != workDir {
-				continue
-			}
-			if got, ok := record["exit_code"].(float64); !ok || int(got) != exitCode {
-				continue
-			}
-			found = record
-		}
-		return nil
-	})
-	if err != nil {
-		t.Fatalf("scan shared runtime logs: %v", err)
-	}
-	if found == nil {
-		t.Fatalf("shared runtime logs contain no %s record for %q exit %d", commandRunnerCompletedLogEvent, workDir, exitCode)
-	}
-	return found
-}
-
-// Fixture returns the root process that owns this scenario.
-func (scenario *Scenario) Fixture() *ProcessFixture {
-	if scenario == nil {
-		return nil
-	}
-	return scenario.fixture
-}
-
-// RootDir returns the temporary root containing the shared process artifacts.
-func (fixture *ProcessFixture) RootDir() string {
-	if fixture == nil {
-		return ""
-	}
-	return fixture.rootDir
-}
-
-type baseCaptureCommandRunner struct {
-	mu    sync.Mutex
-	calls int
-}
-
-func (runner *baseCaptureCommandRunner) Run(_ context.Context, _ platformprocess.CommandRequest) (platformprocess.CommandResult, error) {
-	runner.mu.Lock()
-	runner.calls++
-	runner.mu.Unlock()
-	return platformprocess.CommandResult{Stdout: []byte("script-output-ok")}, nil
-}
-
-func (runner *baseCaptureCommandRunner) CallCount() int {
-	runner.mu.Lock()
-	defer runner.mu.Unlock()
-	return runner.calls
-}
-
-type baseFailureCommandRunner struct{ message string }
-
-func (runner baseFailureCommandRunner) Run(_ context.Context, _ platformprocess.CommandRequest) (platformprocess.CommandResult, error) {
-	return platformprocess.CommandResult{Stderr: []byte(runner.message), ExitCode: 1}, nil
-}
-
-type baseTimeoutThenSuccessCommandRunner struct {
-	mu    sync.Mutex
-	calls int
-}
-
-func (runner *baseTimeoutThenSuccessCommandRunner) Run(ctx context.Context, _ platformprocess.CommandRequest) (platformprocess.CommandResult, error) {
-	runner.mu.Lock()
-	runner.calls++
-	call := runner.calls
-	runner.mu.Unlock()
-	if call == 1 {
-		<-ctx.Done()
-		return platformprocess.CommandResult{}, ctx.Err()
-	}
-	return platformprocess.CommandResult{Stdout: []byte("script-output-after-retry")}, nil
-}
-
-func (runner *baseTimeoutThenSuccessCommandRunner) CallCount() int {
-	runner.mu.Lock()
-	defer runner.mu.Unlock()
-	return runner.calls
-}
-
-type baseCanceledCommandRunner struct{}
-
-func (baseCanceledCommandRunner) Run(context.Context, platformprocess.CommandRequest) (platformprocess.CommandResult, error) {
-	return platformprocess.CommandResult{}, context.Canceled
-}
-
-func assertBaseSessionPlaces(t testing.TB, listed factoryapi.ListWorkResponse, wants map[string]int) {
-	t.Helper()
-	for placeID, want := range wants {
-		if got := support.CountWorkAtCustomerState(listed, placeID); got != want {
-			t.Errorf("%s token count = %d, want %d", placeID, got, want)
-		}
-	}
-}
-
-func assertBaseDispatchOutput(t testing.TB, events []factoryapi.FactoryEvent, want string) {
-	t.Helper()
-	for _, event := range events {
-		if event.Type != factoryapi.FactoryEventTypeDispatchResponse {
-			continue
-		}
-		payload, err := event.Payload.AsDispatchResponseEventPayload()
-		if err != nil {
-			t.Fatalf("decode dispatch response: %v", err)
-		}
-		if payload.Output == nil || *payload.Output != want {
-			t.Fatalf("dispatch output = %#v, want %q", payload.Output, want)
-		}
-		return
-	}
-	t.Fatalf("Factory Event history has no dispatch response: %#v", events)
-}
-
-func assertBaseDispatchErrorContains(t testing.TB, events []factoryapi.FactoryEvent, want string) {
-	t.Helper()
-	for _, event := range events {
-		if event.Type != factoryapi.FactoryEventTypeDispatchResponse {
-			continue
-		}
-		payload, err := event.Payload.AsDispatchResponseEventPayload()
-		if err != nil {
-			t.Fatalf("decode dispatch response: %v", err)
-		}
-		if payload.Error == nil || !strings.Contains(*payload.Error, want) {
-			t.Fatalf("dispatch error = %#v, want substring %q", payload.Error, want)
-		}
-		return
-	}
-	t.Fatalf("Factory Event history has no dispatch response: %#v", events)
-}
-
-func TestProvidersSharedProcessTopology(t *testing.T) {
-	fixture := FixtureFor(t)
-	fixture.AssertTopology(t)
-
-	firstDir := testutilCopySharedFixture(t, "script_executor_dir")
-	secondDir := testutilCopySharedFixture(t, "script_executor_dir")
-	testutil.WriteSeedFile(t, firstDir, "task", []byte("first shared payload"))
-	testutil.WriteSeedFile(t, secondDir, "task", []byte("second shared payload"))
-	first := fixture.OpenScenario(t, firstDir, firstDir, support.NewStaticSuccessCommandRunner("first-shared-output"))
-	second := fixture.OpenScenario(t, secondDir, secondDir, support.NewStaticSuccessCommandRunner("second-shared-output"))
-	if got := fixture.router.routeCount(); got != 2 {
-		t.Fatalf("shared provider active routes = %d, want two", got)
-	}
-
-	var wait sync.WaitGroup
-	wait.Add(2)
-	go func() {
-		defer wait.Done()
-		first.WaitForTerminal(t, FixtureTimeout)
-	}()
-	go func() {
-		defer wait.Done()
-		second.WaitForTerminal(t, FixtureTimeout)
-	}()
-	wait.Wait()
-	firstWork := first.ListWork(t)
-	secondWork := second.ListWork(t)
-	assertBaseSessionPlaces(t, firstWork, map[string]int{"task:done": 1, "task:init": 0})
-	assertBaseSessionPlaces(t, secondWork, map[string]int{"task:done": 1, "task:init": 0})
-	assertBaseDispatchOutput(t, first.FactoryEvents(t), "first-shared-output")
-	assertBaseDispatchOutput(t, second.FactoryEvents(t), "second-shared-output")
-	first.Stop(t)
-	second.Stop(t)
-	if got := fixture.router.routeCount(); got != 0 {
-		t.Fatalf("shared provider routes after cleanup = %d, want zero", got)
-	}
-	fixture.AssertSessionTopology(t)
-}
-
-func TestProvidersSharedProcessRoutes(t *testing.T) {
-	fixture := FixtureFor(t)
-	fixture.AssertTopology(t)
-	baselineCalls := fixture.router.callCount()
-
-	workDir := filepath.Join(fixture.rootDir, "route-test-workdir")
-	routeSelector := fmt.Sprintf("providers-shared-route-test-%d", sharedProviderRouteSequence.Add(1))
-	runner := support.NewStaticSuccessCommandRunner("registered-route-output")
-	if err := fixture.router.register(routeSelector, workDir, runner); err != nil {
-		t.Fatalf("register shared provider test route: %v", err)
-	}
-	defer func() {
-		if err := fixture.router.unregister(routeSelector); err != nil {
-			t.Errorf("unregister shared provider test route: %v", err)
-		}
-	}()
-
-	if err := fixture.router.register(routeSelector, workDir, support.NewStaticSuccessCommandRunner("duplicate-output")); err == nil {
-		t.Fatal("duplicate shared provider route registration succeeded")
-	}
-	if got := fixture.router.routeCount(); got != 1 {
-		t.Fatalf("shared provider route count after duplicate registration = %d, want one", got)
-	}
-
-	result, err := fixture.router.Run(context.Background(), platformprocess.CommandRequest{
-		Command: "echo", WorkDir: workDir, Stdin: []byte("registered route input"),
-	})
-	if err != nil {
-		t.Fatalf("registered shared provider route: %v", err)
-	}
-	if string(result.Stdout) != "registered-route-output" {
-		t.Fatalf("registered shared provider route output = %q, want registered output", result.Stdout)
-	}
-	if got := fixture.router.callCount(); got != baselineCalls+1 {
-		t.Fatalf("shared provider route calls after registered route = %d, want %d", got, baselineCalls+1)
-	}
-
-	unknownWorkDir := filepath.Join(fixture.rootDir, "unknown-route-workdir")
-	_, err = fixture.router.Run(context.Background(), platformprocess.CommandRequest{
-		Command: "echo", WorkDir: unknownWorkDir, Stdin: []byte("must not cross a route"),
-	})
-	if err == nil || !strings.Contains(err.Error(), "no provider route matched WorkDir") {
-		t.Fatalf("unknown shared provider route error = %v, want explicit route failure", err)
-	}
-	if got := fixture.router.callCount(); got != baselineCalls+1 {
-		t.Fatalf("shared provider route calls after unknown route = %d, want %d", got, baselineCalls+1)
-	}
-}
-
-func TestProvidersSharedProcessAdverseRecovery(t *testing.T) {
-	fixture := FixtureFor(t)
-
-	t.Run("invalid_template", func(t *testing.T) {
-		dir := testutilCopySharedFixture(t, "script_executor_dir")
-		writeBaseFixtureFile(t, dir, []string{"workstations", "run-script", "AGENTS.md"}, "---\ntype: MODEL_WORKSTATION\n---\n{{")
-		testutil.WriteSeedFile(t, dir, "task", []byte("invalid-template-payload"))
-		runner := &baseCaptureCommandRunner{}
-		scenario, listed := RunFactory(t, dir, dir, runner, 5*time.Second)
-		assertBaseSessionPlaces(t, listed, map[string]int{"task:failed": 1, "task:init": 0, "task:done": 0})
-		if got := runner.CallCount(); got != 0 {
-			t.Fatalf("invalid-template provider calls = %d, want zero", got)
-		}
-		assertBaseDispatchErrorContains(t, scenario.FactoryEvents(t), "prompt render failed")
-		scenario.Stop(t)
-	})
-
-	t.Run("dependency_failure", func(t *testing.T) {
-		dir := testutilCopySharedFixture(t, "script_executor_dir")
-		testutil.WriteSeedFile(t, dir, "task", []byte("dependency-failure-payload"))
-		scenario, listed := RunFactory(t, dir, dir, baseFailureCommandRunner{"adverse dependency failure"}, 5*time.Second)
-		assertBaseSessionPlaces(t, listed, map[string]int{"task:failed": 1, "task:init": 0, "task:done": 0})
-		assertBaseDispatchErrorContains(t, scenario.FactoryEvents(t), "adverse dependency failure")
-		scenario.Stop(t)
-	})
-
-	t.Run("timeout", func(t *testing.T) {
-		dir := testutilCopySharedFixture(t, "script_executor_dir")
-		support.WriteWorkstationConfig(t, dir, "run-script", "---\ntype: MODEL_WORKSTATION\nlimits:\n  maxExecutionTime: 10ms\n---\nExecute the script.\n")
-		testutil.WriteSeedFile(t, dir, "task", []byte("timeout-payload"))
-		runner := &baseTimeoutThenSuccessCommandRunner{}
-		scenario, listed := RunFactory(t, dir, dir, runner, 10*time.Second)
-		assertBaseSessionPlaces(t, listed, map[string]int{"task:done": 1, "task:init": 0, "task:failed": 0})
-		if got := runner.CallCount(); got < 2 {
-			t.Fatalf("timeout recovery provider calls = %d, want at least two", got)
-		}
-		scenario.Stop(t)
-	})
-
-	t.Run("cancellation", func(t *testing.T) {
-		dir := testutilCopySharedFixture(t, "script_executor_dir")
-		testutil.WriteSeedFile(t, dir, "task", []byte("cancellation-payload"))
-		scenario, listed := RunFactory(t, dir, dir, baseCanceledCommandRunner{}, 5*time.Second)
-		assertBaseSessionPlaces(t, listed, map[string]int{"task:failed": 1, "task:init": 0, "task:done": 0})
-		assertBaseDispatchErrorContains(t, scenario.FactoryEvents(t), "execution cancelled: context canceled")
-		scenario.Stop(t)
-	})
-
-	t.Run("unknown_route", func(t *testing.T) {
-		dir := testutilCopySharedFixture(t, "script_executor_dir")
-		testutil.WriteSeedFile(t, dir, "task", []byte("unknown-route-payload"))
-		scenario := fixture.OpenScenario(t, dir, "", nil)
-		scenario.WaitForTerminal(t, 5*time.Second)
-		listed := scenario.ListWork(t)
-		assertBaseSessionPlaces(t, listed, map[string]int{"task:failed": 1, "task:init": 0, "task:done": 0})
-		assertBaseDispatchErrorContains(t, scenario.FactoryEvents(t), "script command execution failed")
-		scenario.Stop(t)
-	})
-
-	t.Run("known_good_after_adverse_cases", func(t *testing.T) {
-		dir := testutilCopySharedFixture(t, "script_executor_dir")
-		testutil.WriteSeedFile(t, dir, "task", []byte("known-good-payload"))
-		scenario, listed := RunFactory(t, dir, dir, support.NewStaticSuccessCommandRunner("known-good-output"), 5*time.Second)
-		assertBaseSessionPlaces(t, listed, map[string]int{"task:done": 1, "task:init": 0, "task:failed": 0})
-		assertBaseDispatchOutput(t, scenario.FactoryEvents(t), "known-good-output")
-		scenario.Stop(t)
-	})
-
-	if got := fixture.router.routeCount(); got != 0 {
-		t.Fatalf("shared provider routes after adverse recovery = %d, want zero", got)
-	}
-}
-
-func testutilCopySharedFixture(t *testing.T, name string) string {
-	t.Helper()
-	source := support.LegacyFixtureDir(t, name)
-	return testutil.CopyFixtureDir(t, source)
-}
-
-func writeBaseFixtureFile(t *testing.T, dir string, pathParts []string, content string) {
-	t.Helper()
-	path := filepath.Join(append([]string{dir}, pathParts...)...)
-	if err := os.WriteFile(path, []byte(content), 0o644); err != nil {
-		t.Fatalf("write %s: %v", path, err)
-	}
-}
-
-const (
-	forcedProviderCleanupChildEnv  = "YOU_PROVIDERS_FORCED_CLEANUP_CHILD"
-	forcedProviderCleanupReportEnv = "YOU_PROVIDERS_FORCED_CLEANUP_REPORT"
-)
-
-// TestProviders_ForcedAssertionFailureCleansOwnedResources proves that the
-// package-local process/session/router cleanup callbacks still run when a
-// child test exits through an assertion failure. The child boundary is
-// required because a parent test cannot intentionally fail itself and then
-// inspect its own t.Cleanup results.
-func TestProviders_ForcedAssertionFailureCleansOwnedResources(t *testing.T) {
-	if os.Getenv(forcedProviderCleanupChildEnv) == "1" {
-		runForcedProviderCleanupChild(t)
-		return
-	}
-	runForcedProviderCleanupParent(t)
-}
-
-func runForcedProviderCleanupParent(t *testing.T) {
-	t.Helper()
-
-	reportPath := filepath.Join(t.TempDir(), "forced-provider-cleanup.json")
-	command := exec.Command(os.Args[0], "-test.run=^TestProviders_ForcedAssertionFailureCleansOwnedResources$")
-	command.Env = append(os.Environ(),
-		forcedProviderCleanupChildEnv+"=1",
-		forcedProviderCleanupReportEnv+"="+reportPath,
-	)
-	output, err := command.CombinedOutput()
-	if err == nil {
-		t.Fatalf("forced cleanup child exited successfully; output=%q", output)
-	}
-	if command.Process == nil || command.ProcessState == nil || !command.ProcessState.Exited() {
-		t.Fatalf("forced cleanup child did not exit; error=%v output=%q", err, output)
-	}
-	if command.ProcessState.ExitCode() == 0 {
-		t.Fatalf("forced cleanup child exit code = 0; output=%q", output)
-	}
-
-	report := readForcedProviderCleanupReport(t, reportPath, output)
-	assertForcedProviderCleanupReport(t, report, command.Process.Pid)
-}
-
-func readForcedProviderCleanupReport(
-	t *testing.T,
-	path string,
-	childOutput []byte,
-) forcedProviderCleanupReport {
-	t.Helper()
-
-	payload, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("read forced cleanup report %q: %v; child output=%q", path, err, childOutput)
-	}
-	var report forcedProviderCleanupReport
-	if err := json.Unmarshal(payload, &report); err != nil {
-		t.Fatalf("decode forced cleanup report %q: %v; child output=%q", path, err, childOutput)
-	}
-	return report
-}
-
-func assertForcedProviderCleanupReport(
-	t *testing.T,
-	report forcedProviderCleanupReport,
-	childPID int,
-) {
-	t.Helper()
-
-	if report.ApplicationPID != childPID {
-		t.Fatalf("forced cleanup application PID = %d, want child PID %d", report.ApplicationPID, childPID)
-	}
-	if !report.ProcessDone {
-		t.Fatal("forced cleanup Process.Execute did not finish before cleanup report")
-	}
-	if !report.ListenerClosed {
-		t.Fatal("forced cleanup listener remained reachable after process close")
-	}
-	if len(report.OpenedSessionIDs) != 1 || !reflect.DeepEqual(report.OpenedSessionIDs, report.DeletedSessionIDs) {
-		t.Fatalf("forced cleanup sessions opened=%v deleted=%v, want one deleted opened session", report.OpenedSessionIDs, report.DeletedSessionIDs)
-	}
-	if report.ActiveRoutes != 0 {
-		t.Fatalf("forced cleanup active routes = %d, want zero", report.ActiveRoutes)
-	}
-	if !report.Paths.RootAbsent || !report.Paths.FactoryAbsent || !report.Paths.WorkDirAbsent ||
-		!report.Paths.ReplayAbsent || !report.Paths.RuntimeLogAbsent || !report.Paths.WorktreeAbsent {
-		t.Fatalf("forced cleanup owned paths remain: %#v", report.Paths)
-	}
-}
-
-func runForcedProviderCleanupChild(t *testing.T) {
-	t.Helper()
-
-	reportPath := strings.TrimSpace(os.Getenv(forcedProviderCleanupReportEnv))
-	if reportPath == "" {
-		t.Fatal("forced cleanup child report path is required")
-	}
-
-	var fixture *ProcessFixture
-	var scenario *Scenario
-	paths := forcedProviderCleanupPaths{}
-	t.Cleanup(func() {
-		if err := writeForcedProviderCleanupReport(reportPath, fixture, scenario, paths); err != nil {
-			t.Errorf("write forced cleanup report: %v", err)
-		}
-	})
-
-	var err error
-	fixture, err = newSharedProviderProcessFixture(t)
-	if err != nil {
-		t.Fatalf("build isolated forced-cleanup fixture: %v", err)
-	}
-	t.Cleanup(func() {
-		if err := fixture.close(); err != nil {
-			t.Errorf("close isolated forced-cleanup fixture: %v", err)
-		}
-	})
-
-	paths = prepareForcedProviderCleanupPaths(t, fixture)
-	scenario = fixture.OpenScenario(
-		t,
-		paths.Factory,
-		paths.WorkDir,
-		support.NewStaticSuccessCommandRunner("forced-cleanup-output"),
-	)
-	t.Fatal("intentional assertion failure after acquiring process, session, route, and owned paths")
-}
-
-type forcedProviderCleanupPaths struct {
-	Root       string
-	Factory    string
-	WorkDir    string
-	Replay     string
-	RuntimeLog string
-	Worktree   string
-}
-
-func prepareForcedProviderCleanupPaths(
-	t *testing.T,
-	fixture *ProcessFixture,
-) forcedProviderCleanupPaths {
-	t.Helper()
-
-	factoryDir := filepath.Join(fixture.rootDir, "forced-assertion-factory")
-	if err := copySharedProviderDirectory(
-		support.LegacyFixtureDir(t, "script_executor_dir"),
-		factoryDir,
-	); err != nil {
-		t.Fatalf("copy forced-cleanup Factory: %v", err)
-	}
-	worktree := filepath.Join(factoryDir, "worktree")
-	workDir := filepath.Join(factoryDir, "workdir")
-	for _, path := range []string{worktree, workDir} {
-		if err := os.MkdirAll(path, 0o755); err != nil {
-			t.Fatalf("create forced-cleanup path %q: %v", path, err)
-		}
-	}
-	replayPath := filepath.Join(fixture.rootDir, "forced-assertion.replay.json")
-	runtimeLogPath := filepath.Join(fixture.rootDir, "forced-assertion.runtime.log")
-	for _, path := range []string{replayPath, runtimeLogPath} {
-		if err := os.WriteFile(path, []byte("owned test artifact"), 0o600); err != nil {
-			t.Fatalf("create forced-cleanup artifact %q: %v", path, err)
-		}
-	}
-	return forcedProviderCleanupPaths{
-		Root: fixture.rootDir, Factory: factoryDir, WorkDir: workDir,
-		Replay: replayPath, RuntimeLog: runtimeLogPath, Worktree: worktree,
-	}
-}
-
-type forcedProviderCleanupReport struct {
-	ApplicationPID    int                             `json:"application_pid"`
-	ProcessDone       bool                            `json:"process_done"`
-	ListenerClosed    bool                            `json:"listener_closed"`
-	OpenedSessionIDs  []string                        `json:"opened_session_ids"`
-	DeletedSessionIDs []string                        `json:"deleted_session_ids"`
-	ActiveRoutes      int                             `json:"active_routes"`
-	Paths             forcedProviderCleanupPathReport `json:"paths"`
-}
-
-type forcedProviderCleanupPathReport struct {
-	RootAbsent       bool `json:"root_absent"`
-	FactoryAbsent    bool `json:"factory_absent"`
-	WorkDirAbsent    bool `json:"workdir_absent"`
-	ReplayAbsent     bool `json:"replay_absent"`
-	RuntimeLogAbsent bool `json:"runtime_log_absent"`
-	WorktreeAbsent   bool `json:"worktree_absent"`
-}
-
-func writeForcedProviderCleanupReport(
-	path string,
-	fixture *ProcessFixture,
-	scenario *Scenario,
-	paths forcedProviderCleanupPaths,
-) error {
-	if scenario == nil {
-		return fmt.Errorf("forced cleanup scenario was not opened")
-	}
-	opened, deleted := forcedProviderSessionIDs(fixture)
-	report := forcedProviderCleanupReport{
-		ApplicationPID:    os.Getpid(),
-		ProcessDone:       forcedProviderChannelClosed(fixtureDone(fixture)),
-		ListenerClosed:    forcedProviderListenerClosed(fixtureBaseURL(fixture)),
-		OpenedSessionIDs:  opened,
-		DeletedSessionIDs: deleted,
-		ActiveRoutes:      forcedProviderRouteCount(fixture),
-		Paths: forcedProviderCleanupPathReport{
-			RootAbsent:       pathAbsent(paths.Root),
-			FactoryAbsent:    pathAbsent(paths.Factory),
-			WorkDirAbsent:    pathAbsent(paths.WorkDir),
-			ReplayAbsent:     pathAbsent(paths.Replay),
-			RuntimeLogAbsent: pathAbsent(paths.RuntimeLog),
-			WorktreeAbsent:   pathAbsent(paths.Worktree),
-		},
-	}
-	payload, err := json.MarshalIndent(report, "", "  ")
-	if err != nil {
-		return fmt.Errorf("marshal report: %w", err)
-	}
-	if err := os.WriteFile(path, payload, 0o600); err != nil {
-		return fmt.Errorf("write report: %w", err)
-	}
-	return nil
-}
-
-func forcedProviderSessionIDs(fixture *ProcessFixture) ([]string, []string) {
-	if fixture == nil {
-		return nil, nil
-	}
-	fixture.sessionMu.Lock()
-	defer fixture.sessionMu.Unlock()
-	return append([]string(nil), fixture.openedSessionIDs...), append([]string(nil), fixture.deletedSessionIDs...)
-}
-
-func fixtureDone(fixture *ProcessFixture) <-chan struct{} {
-	if fixture == nil {
-		closed := make(chan struct{})
-		close(closed)
-		return closed
-	}
-	return fixture.done
-}
-
-func fixtureBaseURL(fixture *ProcessFixture) string {
-	if fixture == nil {
-		return ""
-	}
-	return fixture.baseURL
-}
-
-func forcedProviderRouteCount(fixture *ProcessFixture) int {
-	if fixture == nil || fixture.router == nil {
-		return 0
-	}
-	return fixture.router.routeCount()
-}
-
-func forcedProviderChannelClosed(done <-chan struct{}) bool {
-	select {
-	case <-done:
-		return true
-	default:
-		return false
-	}
-}
-
-func forcedProviderListenerClosed(baseURL string) bool {
-	if strings.TrimSpace(baseURL) == "" {
-		return false
-	}
-	// The fixture's shutdown signal proves Process.Execute returned; this
-	// bounded HTTP probe additionally proves the public listener is no longer
-	// reachable after the server's close path completed.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-	request, err := http.NewRequestWithContext(
-		ctx,
-		http.MethodGet,
-		strings.TrimSuffix(baseURL, "/")+"/status",
-		nil,
-	)
-	if err != nil {
-		return false
-	}
-	response, err := http.DefaultClient.Do(request)
-	if err != nil {
-		return true
-	}
-	response.Body.Close()
-	return false
-}
-
-func pathAbsent(path string) bool {
-	_, err := os.Stat(path)
-	return os.IsNotExist(err)
 }
